Refactor MainActivity: now we have a AuthorizationMainActivity that checks user is logged in
