--- android-app/app/src/main/res/layout/activity_display_message.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="tallerii.stories.DisplayMessageActivity">
-
-    <TextView
-        android:id="@+id/labelView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="64dp"
-        android:text="Logueado como "
-        android:textSize="12sp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/usernameView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="28dp"
-        android:text="TextView"
-        android:textAppearance="@android:style/TextAppearance.Material.Display3"
-        android:textColor="@android:color/holo_red_dark"
-        app:layout_constraintBaseline_toBaselineOf="@+id/labelView"
-        app:layout_constraintStart_toEndOf="@+id/labelView" />
-</android.support.constraint.ConstraintLayout>
